Provider filter: move watch provider updating to StreamingSearch
